definition: Optimize compilation
The current code looks like:

```
NewEmojis(
  NewEmoji(...),
  NewEmoji(...),
  [...]
)
```

For reasons I do not fully understand, the Golang compiler makes a mess of
this. It seems like it:

 * calls each `NewEmoji()` and stores the result on the stack;
 * allocates the `[]Emoji` on the heap (because it leaks inside
   `NewEmojis()`;
 * then copies each emoji from the stack to the array on the heap
   one by one.

This results into an absolutely silly binary size. The trivial example
from the README is 4.1MB without the emoji extension (on AMD64), but 12MB
with the emoji extension.

Directly definining the `Emoji` struct decreases this size to 4.6MB,
a more than 15x gain in binary size.